blink: Fix subframe paint holding for ViewTransitions.
During a ViewTransition for a local subframe, we cache the entire iframe
as a screenshot so it can be used to paint the old content while the
update callback is running to update the DOM to the new state.

The cached content is painted using a ViewTransitionContentLayer. This
layer has a bit to indicate whether its content is in the current frame
(live) or not. The content is live in the same frame where the capture
directive is processed. In the next frame, Blink clears the paint
because the iframe is throttled so it's no longer live.

However, we weren't updating the liveness bit on
ViewTransitionContentLayer above until after we get an ack from the GPU
process for the capture directive. This meant that any frames pushed by
Blink between the capture directive and its ack would add this layer in
an incorrect state, causing CC to skip drawing the layer's content.

Fix this by correcting the time for when the liveness state on the layer
is changed.

Fixed: 356643991
Change-Id: I7af7330748dad772653c012c55f460701c2c1b90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5758282
Commit-Queue: Vladimir Levin <vmpstr@chromium.org>
Auto-Submit: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1336385}
NOKEYCHECK=True
GitOrigin-RevId: 326d56208deb4d3c44a4e4a2ae53e22b046f88f6