[PEPC] Add different reasons for element is occluded vs out of viewport
Sites need to do different things depending on IntersectionObserver
tells us why an element isn't visible. If the element is just out of the
viewport or clipped, that's expected and they don't need to do anything.
But if the element is completely in the viewport but is occluded by
something else, that means something went wrong and the site needs to
fix it.

Bug: 347219816
Change-Id: If771128890c3d405d582c9e35f2dbf348d9cdb22
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5749025
Reviewed-by: Yi Gu <yigu@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Thomas Nguyen <tungnh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1336560}
NOKEYCHECK=True
GitOrigin-RevId: c753adad0e650a7291a808b7eb897e2770fc636e