Use the most recent visibility value for on demand visibility requests
This CL updates the tracker to cache the `meets_visibility_threshold_`
value. During on demand visibility requests, this allows us to either:
  * Return the current visibility, if the document is in the
kPaintClean state
  * Return the cached visibility value otherwise

This is needed since `RequestVisibility` will be called when users
switch tab. And when a tab is backgrounded lifecycle updates for that
document are stopped.

Bug: 40275580, b/344946183
Change-Id: I3f3eb1e689a7d4a5e076de7026541682faaa269e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5670552
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Benjamin Keen <bkeen@google.com>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1336711}
NOKEYCHECK=True
GitOrigin-RevId: 6bb153624fe88e60840d60cc49515378eb646803