[EditContext] Use empty control and selection bounds as fallbacks
When there are no valid character bounds,
EditContext::FirstRectForCharacterRange is supposed to fall back to
the selection or control bounds if they have been set. Currently we
use an IsEmpty() check to determine if they've been set, but this
actually returns false if the given rect has zero width or height.

A site may want to set selection to a rect of empty width to indicate
a caret, so this would be skipped as a fallback.

Fix this by instead checking against the default gfx::Rect to
determine whether a selection or control bounds has been set.

Bug: 351029417
Change-Id: I6d519a0e2dab83277b8a5d6df103f0fd8f4a26bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5744458
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Dan Clark <daniec@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1336717}
NOKEYCHECK=True
GitOrigin-RevId: 5d63a287ea479a275bd68a8f9971a1cdde8d9fd9