Fix potential draggable region crash when no MainFrameImpl
This CL fixes a crash that can occur when SetSupportsDraggableRegions
is called with `true` and there is no MainFrameImpl. When MainFrameImpl
is nullptr, logic currently correctly returns early, but supports_draggable_regions_ is set before that happens. As a result, when SupportsDraggableRegions() is called, it will return true, and thus LocalFrameView::UpdateDocumentDraggableRegions() will call DraggableRegionsChanged(). This will trigger a crash in WebViewImpl::DraggableRegionsChanged(), as it assumes that MainFrameImpl is not null.

Chrome likely does not hit this crash because it only collects app regions for Window Controls Overlay & Borderless apps, whereas Electron enables draggable regions for more frames.

Change-Id: I80988734b38f143851043b36914fb5eff6135fd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5756619
Reviewed-by: Amanda Baker <ambake@microsoft.com>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337265}
NOKEYCHECK=True
GitOrigin-RevId: bd4e566885f1830d5f8e8687912c5d87723c075d