[TTF] Introduce likely_contains_captcha in FormData
The autosubmission logic for TTF tried to detect captchas and prevent
autosubmission in this case. After TTF refactoring, the autosubmission
calculation logic was moved into the browser process as opposed to the
renderer. Now it's necessary to pass the knowledge whether the form
contains a captcha to the browser. This CL introduces a new field into
FormData likely_contains_captcha. It is set in
ExtractFormDataWithFieldsAndFrames and then read in password credential
filler.

Bug: 346494850
Change-Id: I90872fd87494848609a17641a60a513cb463067d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5748809
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Christoph Schwering <schwering@google.com>
Commit-Queue: Anna Tsvirchkova <atsvirchkova@google.com>
Cr-Commit-Position: refs/heads/main@{#1337290}
NOKEYCHECK=True
GitOrigin-RevId: fd4e6cbd4e107edca622e6b85f6ae1d1a8deed1f