Change StringHasher into using rapidhash.
StringHasher currently uses SuperFastHash, an algorithm from 2013.
While fairly fast, it does not match the speed of current top hash
functions, and worse, it has a number of statistical problems
(it fails nine out of ten quality tests in SMHasher).

We replace it by rapidhash, the current SMHasher leader in terms of
cycles/hash among those that pass all quality tests (and also way faster
than SuperFastHash in terms of throughput for large inputs). This
gives us a modern (noncryptographic) hash function with good quality,
and increases speed all over. Since we use hashing quite a lot,
in particular during creation of AtomicString, we get a little
speed boost across a fairly large set of benchmarks.

There are some subtleties. SuperFastHash works naturally in terms of
16-bit inputs, so StringHasher would expand all 8-bit (LChar) input
into 16-bit (UChar) before hashing. rapidhash typically reads 64- or
32-bit chunks, so this expansion is less suitable; thus, we generally
prefer to hash the data as raw bytes where we can. However, we still
need to maintain the property that Latin1-only data stored as UChars
to the same values as if stored in LChars.

Furthermore, rapidhash is a larger function than SuperFastHash.
We balance speed and performance by choosing the “compact“ version
(most of our hash inputs are short anyway), and only inlining those
call sites that we've found to be called often during benchmarking.
This keeps bloat down, although there still is some APK size increase
(we stay under warning limit).

We also have another copy of SuperFastHash in //base/hash, but there,
it is used as a persistent hash (i.e., values need to be stable
over time), so it would be impossible to remove it entirely.
It is possible that we should look at replacing CityHash with rapidhash,
though, in order to get fewer hash functions.

Some tests depend on the ordering of hash tables. Most of them
have been fixed in preparatory CLs, but there are some that are
difficult to fix (e.g. devtools tests where the code and the recorded
test results live in different repositories, but need to be rolled
in sync if we change the implementation) that are simply re-recorded
in this changelist. If we ever start randomizing the StringHasher
secret (which is an option now with rapidhash), we'll need to fix
them so that they are deterministic.

Speedometer3 (M1 Pinpoint, no PGO), significant results at 99% only,
lower is better except for score:

  NewsSite-Next                                 [ -1.4%,  -0.2%]
  NewsSite-Nuxt                                 [ -1.6%,  -0.2%]
  Perf-Dashboard                                [ -1.2%,  -0.1%]
  React-Stockcharts-SVG                         [ -1.0%,  -0.0%]
  Score                                         [ +0.1%,  +0.8%]
  TodoMVC-JavaScript-ES5                        [ -1.1%,  -0.2%]
  TodoMVC-JavaScript-ES6-Webpack-Complex-DOM    [ -0.9%,  -0.4%]
  TodoMVC-jQuery                                [ -2.2%,  -1.3%]
  TodoMVC-Lit-Complex-DOM                       [ -1.7%,  -0.3%]

MotionMark (M1 Pinpoint, no PGO), significant results at 99% only,
higher is better:

  Leaves                                        [ +0.2%,  +0.8%]
  motionmark_ramp_composite                     [ +0.3%,  +1.0%]
  Multiply                                      [ +0.9%,  +1.4%]
  Suits                                         [ +1.0%,  +2.5%]

Speedometer3 (Win10 Pinpoint, otherwise same as M1):

  TodoMVC-JavaScript-ES6-Webpack-Complex-DOM    [ -1.5%,  -0.1%]
  TodoMVC-Svelte-Complex-DOM                    [ -3.2%,  -0.3%]

Pixel6 Speedometer3 is generally neutral (if anything, slightly
positive).

Style parsing perftest (Zen 3, LTO but no PGO):

  Parse (µs)             Before     After    Perf      95% CI (BCa)
  =================== ========= ========= ======= =================
  ECommerce                1016       989   +2.8%  [ +1.8%,  +3.7%]
  Encyclopedia             5713      5467   +4.5%  [ +3.6%,  +5.4%]
  Extension                1019       986   +3.4%  [ +2.2%,  +4.5%]
  News                     6097      5889   +3.5%  [ +2.6%,  +4.4%]
  Search                   3737      3663   +2.0%  [ +1.1%,  +2.9%]
  Social1                 10903     10553   +3.3%  [ +2.4%,  +4.2%]
  Social2                   463       452   +2.5%  [ +1.3%,  +3.5%]
  Sports                  41492     40131   +3.4%  [ +2.4%,  +4.3%]
  Video                   23832     22934   +3.9%  [ +2.1%,  +5.6%]
  Geometric mean                            +3.3%  [ +2.4%,  +4.1%]

I've also spot-checked that we do OK in PGO runs (actually slightly
better than in the non-PGO runs), even though we should be at a
disadvantage due to all the rapidhash code having no PGO coverage
whereas the SuperFastHash code has.

Change-Id: I3f27da4cb2df4ed7120306530b56e83a65d01886
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5667077
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337378}
NOKEYCHECK=True
GitOrigin-RevId: 98742ca1b98b0598b1981671abb994c1a442ba6e