Add transferability to RTCDataChannel
This CL adds all necessary methods to allow an RTCDataChannel to be
transferred. Notably, it delays registering the RTCDataChannel's
observer, and handles state clean-up after a transfer, according to
the spec [1].

Delaying the registration of `observer_` gives a small window of time
for a Window to transfer an RTCDataChannel to a worker without losing
any messages.

As soon as a blink::RTCDataChannel's observer is registered, it drains
the incoming message queue from the underlying webrtc::DataChannel,
posting messages from the webrtc network thread to the main thread.
These incoming messages could be lost if the RTCDataChannel was
transferred when the messages are in flight.

Note: transferring RTCDataChannels will be enabled in an upcoming CL.
See https://chromestatus.com/feature/6612726584180736.

[1] https://w3c.github.io/webrtc-extensions/#rtcdatachannel-transferable

Bug: 40787712
Change-Id: I8da7a3878ee9dd0a8c46d9681d756d4a9e6e8c5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5751286
Reviewed-by: Tony Herre <toprice@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Florent Castelli <orphis@chromium.org>
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337402}
NOKEYCHECK=True
GitOrigin-RevId: 6e64019985982bf535d397aabd5865fda0197af7