[performance] Use V8 fast API for performance.now()
The fast API can be used for performance.now() now, even though there is
no performance improvement yet. With this change, the HeapNumber for the
result value gets allocated in V8 generated code, and not in the
bindings code anymore. V8 can be optimized though to avoid the
HeapNumber allocation, which is a bit difficult though at the moment for
architectural reasons.

Bug: 41497260
Change-Id: I3a7b6a59e2b8a0a8a90f578b1320a285cdc4d08d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5756588
Reviewed-by: Annie Sullivan <sullivan@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337674}
NOKEYCHECK=True
GitOrigin-RevId: 7a0b3471d06e18d9f103a1a9acd338f01a17d31a