Fix direction: rtl for scroll markers tracking
If the element's direction is rtl (right-to-left) and content grows to the
left, then scrollLeft is 0 when the scrollbar is at its rightmost position
(at the start of the scrolled content), and then increasingly negative
as you scroll towards the end of the content. So we can just use abs
function to cover both cases.

Bug: 332396355
Change-Id: I583f01dacdeed58b09958ce367b06f1fa3aa56c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5760163
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Daniil Sakhapov <sakhapov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1337827}
NOKEYCHECK=True
GitOrigin-RevId: 2021d3ffca21b943144918c621348c0bcaa4523b