[EditContext] Check whether frame exists in SetSelection
EditContext.updateSelection can be called even when the EditContext
is in a detached frame, causing a crash because SetSelection assumes
that DomWindow()->GetFrame() is always non-null.

Fix the crash by adding a null-check.

Note that EditContext has other usages of DomWindow()->GetFrame() that
are not protected by a null-check. As far as I can tell these other
usages should not occur in a detached frame. If however we start
seeing crashes around these other usages, we should consider
protecting all of them with similar checks.

Bug: 356211970
Change-Id: I586be30f820ee8470f5d3f5c8c1c5731ca14e558
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5761754
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Dan Clark <daniec@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1337926}
NOKEYCHECK=True
GitOrigin-RevId: 8dd35520c00bafc9001ffe4cd514b32ddb2e1884