Scheduling APIs: Remove scheduler.yield() parameters
By default, scheduler.yield() continuations inherit the priority and
abort signal from the current task, but this behavior can be overridden
by providing a signal or priority option. But it's not clear if this is
necessary, and there are questions around the expectations of subsequent
continuations (see https://github.com/WICG/scheduling-apis/issues/96).

So to simplify the API and prevent shipping behavior we might regret,
we're removing the yield() parameters. If control over a continuation's
priority/abort is needed, a new task can be scheduled using postTask().
We plan to reevaluate this as the API rolls out and developers start
using it in more complex ways.

Related PR: https://github.com/WICG/scheduling-apis/pull/100.

Bug: 40633887
Change-Id: I5e23c9c8cf950cec795cddafd27c962b71eea5f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5753325
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1338805}
NOKEYCHECK=True
GitOrigin-RevId: ea2bd51cd50579af2273f3daa93053c001bd6cef