Add a WPT test validating that reset after flush discards frame content
There was already a test validating that reset clears the frame buffer
(2d.reset.basic), but that test called reset right after the draw
command, in the same frame. Implementation might behave differently if
a frame is actually rasterized before reset is called.

In particular, Chromium has an optimization that discards pending paint
ops when reset is called. Thus, in the original test, the frame buffer
was never actually written to before the reset and the test could not
verify whether reset actually clears that buffer. By forcing frame to
be rasterized, we now validate that the frame buffer is correctly
cleared.

Bug: 354124879
Change-Id: Ia6536568e2666138ffa4a22418a7c4eb14983577
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5746932
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Yi Xu <yiyix@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1339041}
NOKEYCHECK=True
GitOrigin-RevId: fd68058742e83ab66ca44a20c59dcc94fbc50ae3