Added Reporting changes to allow for deal support.
Originally, reportWin only took one form of reporting id.
This change does not modify the behavior when
selectedBuyerAndSellerReportingId is not provided. However,
when selectedBuyerAndSellerReportingId is provided, we allow for
buyerReportingId and buyerAndSellerReportingId to be passed into
reportWin along with it. This also changes reportResult, but more
trivially. Currently, if buyerAndSellerReportingId is present, it's
provided to reportResult; with this change, if
selectedBuyerAndSellerReportingId is present, it too is provided
to reportResult.

Also added a new enum to the
Ads.InterestGroup.{Server}Auction.ReportingIdType, that will show if
selected_buyer_and_seller_reporting_id is present.

More context on the deals can be found at
https://github.com/WICG/turtledove/issues/873.

Bug: 356654297
Change-Id: I3370e3887db3fe1a18d82246a5f21a78237bff34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5746492
Reviewed-by: Orr Bernstein <orrb@google.com>
Commit-Queue: Youssef Bourouphael <ybourouphael@google.com>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1339099}
NOKEYCHECK=True
GitOrigin-RevId: 99854cd12c8046ac79637cd8b8e46465c46e78a9