Implement autofill preview for appearance:base-select
This patch migrates the autofill popover for <selectlist> to
<select style="appearance:base-select"> because appearance:base-select
is replacing <selectlist>.

I also added some CSS to make the popover take up the same space as the
select's button and have autofill colors.

There are no tests for the visual appearance of form control elements
when they are in the preview state at all, so we will have to wait to
test the visual appearance of this popover until that is possible.

Bug: 40258765, 335457051, 351971330
Change-Id: I5a14f3bca69e40f612b698beee307787d5cd1738
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5719009
Reviewed-by: Peter Kotwicz <pkotwicz@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1339246}
NOKEYCHECK=True
GitOrigin-RevId: 97e663549210702b69370bc6319b9f114667385b