Move EventData information to PerformanceEventTiming
Currently we store some information about Event Timing inside a PerformanceEventTiming, and some more information in an internal EventData object. Much of this data is only used temporarily or for internal reporting and is discarded by the time Entry is added to the performance timeline. We already weren't consistent and some internal-only data is stored in both places.

This change moves the EventData object to be stored within the PerformanceEventTimingEntry object, but makes sure it is discarded before reporting to Event Timing API.

This makes the code cleaner and safer.

Change-Id: I738d209038c5557dccebd4c9a03ba579a839f853
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5692440
Commit-Queue: Hao Liu <haoliuk@chromium.org>
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1339253}
NOKEYCHECK=True
GitOrigin-RevId: b1e6927d0f8a8895994c11c47a2c4e71b6796914