Fix :active state applying to control during second click on <label>
Prior to this change, clicking the <label> of a focused button, checkbox or radio would appear to not apply :active state to the element during a mousedown. Because of this, clicking on a <label> twice (without double clicking) would seem to not apply :active state to its control on the second click because the first click had focused the control.

The reason a focused control appeared to not gain :active state was because performing a mousedown on the label would blur the control, which cleared :active state right after it was applied by the label code.

This fix stops button, checkbox and radio elements from clearing their :active state on blur if they have an :active label, so that :active state is visibly applied on the second click of the label.

Fixed: 40934455
Change-Id: Ic10fd2b791994c1f98f314af6fa0c66e9266356c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5724697
Reviewed-by: David Baron <dbaron@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1339868}
NOKEYCHECK=True
GitOrigin-RevId: bdb93badc1c23030f6693b3308ceb85cef7972e9