Allow anchor attribute to be used on Element, not just HTMLElement
When [1] landed, most of the anchor attribute handling was
moved from HTMLElement to Element. However, the code to watch
anchor references (in AnchorElementObserver) was not similarly
updated. That led to crashes.

This moves the rest of the anchor element stuff to Element,
and adds a test.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/5229437

Fixed: 353865123
Bug: 342579983
Change-Id: I1151c82dfcb39e7ca2db85e7374903e07b113a9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5777660
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Mason Freed <masonf@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1339932}
NOKEYCHECK=True
GitOrigin-RevId: 716e5d1c26b459d2a4166c498668f50c38292609