Remove some of the tiny functions in CSSTokenizer.
Back when we had a jump table in CSSTokenizer, we needed to have one
microfunction per possible character class. We don't anymore, but they
remain, many of them with a 'cc' parameter that isn't really used
for anything. Manually inline the ones that are small enough that
the code becomes clearer.

This shouldn't really affect the generated code; these would be
inlined with or without the change.

Change-Id: I3b55df8aa63b0f779cf784b6549118dacb9d5acd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5781393
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1340330}
NOKEYCHECK=True
GitOrigin-RevId: b63197d87853e23af84ba00745b8e80338e434aa