Ignore vs during system fallback on various platforms
Chrome is using platform specific implementation of
`PlatformFallbackFontForCharacter` to get fallback font on each
platform. It relies on FontFallbackPriority value in order to retrieve
a font with the correct emoji presentation.

Emoji System Fonts on Mac, Win and Android either do not have cmap 14
subtable or it does not include all emojis from their cmap table. We use
cmap 14 subtable to identify whether there is a colored (emoji
presentation) or a monochromatic (text presentation) glyph in the font.
This may lead to the cases when we will not be able to get the glyph ID
for the requested variation sequence using fallback system font and will
continue the second shaping fallback list pass ignoring variation
selectors, see [0]. For example:

<div style=”font-family:web-mono-font”>&#x1FAE8;&#xFE0F</div>

if "web-mono-font" has a glyph for U+1FAE8 codepoint, we will use this
monochromatic web font instead of using system colored font.

To prevent that once we reached system fallback fonts, we can ignore
emoji variation selectors since we will get the font with the correct
presentation relying on FontFallbackPriority.

[0] https://docs.google.com/document/d/1NyIKm0PnWUwX6j0smDwxDAPSUoiyBHPL95oH2lvjjpc

Bug: 40628044
Change-Id: If0195171ab7f76f10a6f65ad686c2fd8427c0502
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5771131
Commit-Queue: Munira Tursunova <moonira@google.com>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1340368}
NOKEYCHECK=True
GitOrigin-RevId: d95ffd45a497f59cc88edc7c88af0aff6bf261a7