Include background image of image-set as LCP candidates
This is to include background images specified by the image-set function
as LCP candidates.

The parameter type StyleFetchedImage is changed to StyleImage in the
LCP code to accommodate this change. The latter is a parent class of the
former and the StyleImageSet class which corresponds to the background
images specified by the image-set function.

Bug: 40929338
Change-Id: I1725cc9beab06ff17e44e4f82b26ab7f582191d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5645790
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Hao Liu <haoliuk@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1340668}
NOKEYCHECK=True
GitOrigin-RevId: 9f280229a11c07ab05ed057a71074ee0d0786786