[calc-size] Allow applying calc-size to the auto min-size.
This patch shifts where we apply the auto min-size. Previously we
change the blink::Length we resolved. This patch passes in an
auto_length to resolve if the original length has auto instead.

Fixes: block-aspect-ratio-038.html as we weren't correctly clamping
the auto-min-size by the max-size in the inline direction.

Bug: 1343829, 313072
Change-Id: Ia56ea5e52b3a42c5b7af8447eb19888883e7e76e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5753047
Reviewed-by: David Baron <dbaron@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1340677}
NOKEYCHECK=True
GitOrigin-RevId: d8dd9b8049a717b947ce40044082b0db8f65574a