Add slow test meta tag to test that gets timed out on buildbots
The test "control-active-with-multiple-clicks.html" added
in a previous change
(https://chromium-review.googlesource.com/c/chromium/src/+/5724697)
is getting timed out on Mac and Win11 buildbots,
which is why the test has been disabled for them.

This fix adds a meta tag to the test indicating that it is slow, so that
it will no longer be timed out.

Fixed: 358591568, 358437643
Bug: 40934455
Change-Id: Ida9cc1e9c4d42b2fd2989af81234ea5e598e6ad7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5782735
Reviewed-by: David Baron <dbaron@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341236}
NOKEYCHECK=True
GitOrigin-RevId: 2b7b2a38454b8eb3cc1c62cd52be6a121711fcb6