Do not show duplicated logs
On Android we are seeing:
dbus is not running, not showing results...
Failed to open bla bla ...

It's confusing to see both lines of log printed out. Reorg the
code so that the second line of code is not printed when dbus
is not running.

Bug: None
Change-Id: I75d6462f141baa5b2b5ab11e073785f4ffa6f698
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5785808
Commit-Queue: Weizhong Xia <weizhong@google.com>
Auto-Submit: Weizhong Xia <weizhong@google.com>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341322}
NOKEYCHECK=True
GitOrigin-RevId: 32641400e99116ece7363eeb44461ab15755998f