[text-box-trim] Rename `text-box-edge` values
This patch renames `text-box-edge` values from:
```
  text-box-edge: leading | <text-edge>
```
to:
```
  text-box-edge: auto | <text-edge>
```

Note the `<text-edge>` definition is unchanged.
```
  <text-edge> = [ text | cap | ex ] [ text | alphabetic ]?
```

as per the resolution at:
https://github.com/w3c/csswg-drafts/issues/10675#issuecomment-2274517610
and the new spec at:
https://drafts.csswg.org/css-inline-3/#propdef-text-box-edge

As per the new spec, the new `auto` value is identical to
`text` when the `line-fit-edge` property is initial.

The functionality the `leading` value provided is removed.

Bug: 358204934, 40254880
Change-Id: I738058cc79b5525b2cbb151b5765f65363b44e7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5776349
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1341434}
NOKEYCHECK=True
GitOrigin-RevId: 20f1c0306ad81042d177ef97e7002226afa97041