webnn: Support triangular in tflite converter
There is no equivalent operator in TFLite to implement the operation,
this CL emulates it with element-wise mul the mask matrix and input.

The mask matrix is generated with the location of element in the matrix
that is the row (i) and column (j), the element is in the upper
triangular if `j >= i + diagonal`, otherwise it is in the lower
triangular.

Bug: 40206287
Change-Id: I53f6a7b2b12d2805a1425a07cc83b0a39d88f072
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5755723
Reviewed-by: ningxin hu <ningxin.hu@intel.com>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Junwei Fu <junwei.fu@intel.com>
Cr-Commit-Position: refs/heads/main@{#1341589}
NOKEYCHECK=True
GitOrigin-RevId: a3e13366450e60d832fba235fd7ce51ed8d1a0ea