Modify k-anon for reporting key to include selected reporting id.
Today, a bid is allowed to win an auction only if it's k-anonymous.
There's a separate "k-anonymity for reporting" that includes the
renderURL and all of the reporting IDs conveyed to the reporting
function as part of the k-anon key. If a winning bid is not
k-anonymous for reporting, the reporting IDs are all omitted in the
call to the reporting functions.

With the change to reporting IDs to support deals, described in
https://github.com/WICG/turtledove/issues/873, a new reporting ID is
added, selectedBuyerAndSellerReportingId, which adds new k-anon for
reporting keys, multiple of which may need to be fetched for each ad
as the selectedBuyerAndSellerReportingId may be any one of the
selectableBuyerAndSellerReportingIds expressed in the InterestGroup ad.

Change-Id: Ife157dfa59e4e6a8fa6421ee87a5238d57110e11
Bug: 334053709
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5782241
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Orr Bernstein <orrb@google.com>
Cr-Commit-Position: refs/heads/main@{#1341902}
NOKEYCHECK=True
GitOrigin-RevId: c8f95be8df7b5b9e449ce4220b1817bd136ba8b0