WebTransport: Don't try to log binary data as string
WebTransport OutgoingStream had two DVLOG() statements where the result
of base::span::data() was passed to operator<<. This results in it being
treated as a null-terminated string, resulting in an out-of-bounds read
if no null is found.

Fix it by casting to void*, so that the pointer is formatted as a
pointer.

BUG=358538899

Change-Id: I7cc5b2cf7699cff34bcde149ba4e34d7861cebaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5782686
Commit-Queue: Adam Rice <ricea@chromium.org>
Reviewed-by: Nidhi Jaju <nidhijaju@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1342123}
NOKEYCHECK=True
GitOrigin-RevId: 44a6971702ccf11e29f5582703c8726be6eefab2