Fixes lost fin in webtransport, if fin arrives before bidi stream
If a server initated stream's reaches the renderer, before the stream's
fin reaches the renderer, the fin is lost.
For server initited stream this is already taken care of, this fix copies the unidirectional code over to the bidirectional case to address this issue.

Bug: 357281124
Change-Id: I82e55844c5ad0c19a152c79c86d563dcddb5c31d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5759788
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Victor Vasiliev <vasilvv@chromium.org>
Reviewed-by: Nidhi Jaju <nidhijaju@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1342193}
NOKEYCHECK=True
GitOrigin-RevId: ec7a03b40446049bab2d732eed70b3ccf96632be