Templatize RuleInvalidationDataVisitor for Builder and Tracer variants
Wrap the spots where we mutate the invalidation data in compile-time
checks so that (a) they're only included in the Builder case; (b) the
Tracer can treat all the data it's looking at as const.

In some cases, calls that were directly on an InvalidationSet are
wrapped in helper methods. This is to facilitate adding tracing
alongside those calls in the Tracer case.

Several places where InvalidationSets were passed by reference also
needed to be switched to pointers. This is because there are cases
where the Tracer may not find an InvalidationSet we expect to be
there. One example is when, at the time we run the Tracer, a rule has
been added to a stylesheet but not yet indexed. In such a case, the
InvalidationSetToSelectorMap will pick up information about the new
rule as it gets indexed on the next document lifecycle update.

Bug: 337076014
Change-Id: I2151f0103628c5279485cf3fb05bd08de50e69d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5722930
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1342338}
NOKEYCHECK=True
GitOrigin-RevId: b9439522925b7f72ac3069275604742167cfb367