[blink] Add LocalMainFrame::Discard()
This CL adds a new method to LocalMainFrame that configures the
LocalMainFrame for a discard operation.

Specifically it replaces the current blink document with a new empty
document. This is done with the intention of clearing away resources
associated with the LocalFrame and its frame tree.

The implementation is intended to mirror closely javascript: URL
functionality that also replaces the current blink document. The
separate implementation for Discard() avoids overloading the
javascript: URL code path and allows eliding unnecessary checks and
updates implemented specifically for javascript: URLs.

There are no uses of this new API in this CL or any behavior changes.
It is used instead in the follow up CL.

See doc below and crrev.com/c/5637909 for details.
https://docs.google.com/document/d/1pwZ-kCD4Zsijyikd7Saok0g2Qdgr8z8FC_kDqX6pTnw/edit?usp=sharing

Bug: 347770670
Change-Id: Id57373c59883d64671b7479d9b6ad9bbeabae9a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5782189
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1342520}
NOKEYCHECK=True
GitOrigin-RevId: 7485b61f5c249957268e748706cb7d16f22266af