androidx: Swap desktop and jvmstubs for android
In lieu of a better way to specify/choose android for transitive deps
instead of desktop/jvmstubs, we can explicitly replace them in target
names assuming that the android equivalent exists. If anyone can figure
out a way to test attributes and choose based on attributes, that would
be a better way.

Added back androidx.savedstate:savedstate dep so that the roller has a
chance of succeeding after this CL lands, as it seems like the roll may
not need any other changes. When the roller is green again then we can
manually remove that dep and the deps entries.

Bug: 359896493
Change-Id: I691a8ed124cf5a4439316dc8e3c738a408e121bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5789638
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Auto-Submit: Peter Wen <wnwen@chromium.org>
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1342618}
NOKEYCHECK=True
GitOrigin-RevId: cae9e131bd7803472258fa79fb0ea9fac005c395