sideways: Fix `InlineNode::SegmentFontOrientation()`
Segmenting for the font orientation is needed only for
vertical typographic mode, not for vertical flow. Computing
runs is harmless as it computes to single run, but a waste.

This patch also adds `IsHorizontalTypographicMode()` to
`LayoutObject` and `InlineNode` to take advantage of the
bitfield in `LayoutObject`.

Bug: 40501131
Change-Id: Id77b254c27df303c5cdbbcfa4763913c779294ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5784250
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Auto-Submit: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1342671}
NOKEYCHECK=True
GitOrigin-RevId: 3d954b4846b0254e37853a34c163c678814a0ec9