[css-nesting] Remove unnecessary skip to kSemicolonToken
Just a minor cleanup: we already know that we are at kSemicolonToken
in this branch, so there's no point in calling
SkipUntilPeekedTypeIs<kSemicolonToken>. There is also no point
in checking AtEnd() before consuming, since it's legal to
consume the kSemicolonToken, even when kSemicolonToken is a boundary
(although it isn't in this case).

There should be no behavior change.

Bug: 343463516
Change-Id: I05485c0a8dddbc6bde6ed2cab8e25b43cc7b1d86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5793115
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1342736}
NOKEYCHECK=True
GitOrigin-RevId: 6b97f0b307e85cf75d2648b0891a3338cbc022ff