Add unit test for allowing PT reuse on different transports.
When stuff is bundled, PTs must be unique across the bundle.
When it is unbundled, the same PT can be used for different types on
different transports.

This test exercises that scenario.

The TestExpectations change is an irrelevant change required for presubmit
to pass.

Bug: None
Change-Id: Id698112e6e3704c437221ea2963fdaf0c863ac04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5786997
Reviewed-by: Florent Castelli <orphis@chromium.org>
Commit-Queue: Harald Alvestrand <hta@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1342762}
NOKEYCHECK=True
GitOrigin-RevId: b6ebaf26c062b15801dc8e3f6e680dbc7f590faf