Apply extent checks in both dimensions for dynamic expansion
This CL restructures the code to apply scroll extents checks in
both dimensions in addition to clipping by the contents rect
(this can be more constraining on the cull rect outsets in some
cases). This CL also fixes an issue preventing outset from one
dimension to be transferred to another in the case that the
scroller can be scrolled in two dimensions.

Bug: None
Change-Id: I63a6b7a20723d66f02ea7cdc327f70891868ab3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5784052
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Ian Vollick <vollick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1342774}
NOKEYCHECK=True
GitOrigin-RevId: 426098dca3a807a184e35d7eaa01e0caefec5d1f