[AccentColor] Plumb `is_in_web_app_scope` to StyleColor::ColorFromKeyword()
To ensure that the `AccentColor` system keyword is correctly applied
only within the scope of installed WebApps, it's necessary to
introduce a boolean flag within
`LayoutTheme::GetAccentColorOrDefault()`. This flag will enable this
method to determine whether the content falls under the WebApp's
domain, as it doesn't inherently have access to the `Document`.

This initial CL will facilitate the passing of this value up to the
`StyleColor::ColorFromKeyword()` function. Subsequent updates will
then integrate this value within the
`LayoutTheme::GetAccentColorOrDefault()` to achieve the
desired behavior.

Bug: 40229450
Change-Id: Ie59f1f29fed7d248d0dd1c71c9b839eda3544384
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5785551
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Sam Davis Omekara <samomekarajr@microsoft.com>
Commit-Queue: Priya Palanisamy <priyapal@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1342935}
NOKEYCHECK=True
GitOrigin-RevId: 00e672f71a4e55096947c48e8ddde0a79bd47c01