Include libxml revision in README.chromium
This patch adds the commit hash as "Revision" in README.chromium. The
commit hash is already being used as the "Version" in README.chromium,
which I plan to fix later since libxml has added a VERSION file in a
commit which we haven't rolled to yet because of issues with those newer
commits.

Fixed: 349530088
Change-Id: Ie58c9466913ff28aeb6fd15f15de7246b7059735
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5685044
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Reviewed-by: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1343075}
NOKEYCHECK=True
GitOrigin-RevId: f5082d2d86a58b62190037be3e149183d5ef7cac